Refactored code to separate class files.
